requirements.txt
asgiref==3.3.4
Django==3.2
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
PyJWT==2.0.1
pytz==2021.1
screen==1.0.1
sqlparse==0.4.1
uWSGI==2.0.19.1
xlwt==1.3.0
